requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
requests==2.32.3
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.1
prophet==1.1.5
plotly==5.24.1
gunicorn==22.0.0
numpy==1.26.4
reportlab

